+++ greenink/dom/__init__.py
@@ -1,0 +1,4 @@
+from .element import HTMLElement
+from .document import HTMLDocument
+
+__all__ = ["HTMLElement", "HTMLDocument"]